+++ src/main/java/ch/zhaw/psit4/services/implementation/ConfigServiceImpl.java
@@ -1,13 +1,8 @@
 package ch.zhaw.psit4.services.implementation;
 
-import ch.zhaw.psit4.data.jpa.repositories.CompanyRepository;
 import ch.zhaw.psit4.data.jpa.repositories.SipClientRepository;
 import ch.zhaw.psit4.domain.ConfigWriter;
 import ch.zhaw.psit4.domain.ConfigZipWriter;
-import ch.zhaw.psit4.domain.beans.Company;
-import ch.zhaw.psit4.domain.beans.DialPlanContext;
-import ch.zhaw.psit4.domain.beans.SipClient;
-import ch.zhaw.psit4.domain.dialplan.helper.ContextGenerator;
 import ch.zhaw.psit4.domain.exceptions.InvalidConfigurationException;
 import ch.zhaw.psit4.domain.exceptions.ZipFileCreationException;
 import ch.zhaw.psit4.domain.interfaces.DialPlanContextConfigurationInterface;
@@ -18,7 +13,6 @@ import ch.zhaw.psit4.services.interfaces.ConfigServiceInterface;
 import org.springframework.stereotype.Service;
 
 import java.io.ByteArrayOutputStream;
-import java.util.ArrayList;
 import java.util.List;
 
 /**
@@ -36,26 +30,9 @@ public class ConfigServiceImpl implements ConfigServiceInterface {
     private final SipClientConfigAdapter sipClientConfigAdapter;
     private final DialPlanConfigAdapter dialPlanConfigAdapter;
 
-    public ConfigServiceImpl(SipClientRepository sipClientRepository, CompanyRepository companyRepository) {
-        this.dialPlanConfigAdapter = new DialPlanConfigAdapter(sipClientRepository, companyRepository);
+    public ConfigServiceImpl(SipClientRepository sipClientRepository) {
+        this.dialPlanConfigAdapter = new DialPlanConfigAdapter(sipClientRepository);
         this.sipClientConfigAdapter = new SipClientConfigAdapter(sipClientRepository);
-    }
-
-    /**
-     * Converts a sip client entity of the data storage into a domain specific sip client.
-     *
-     * @param sipClient the sip client entity of the data storage
-     * @return the domain specific sip client
-     */
-    public static SipClient sipClientEntityToSipClient(ch.zhaw.psit4.data.jpa.entities.SipClient sipClient) {
-        SipClient sipClientDomain = new SipClient();
-        sipClientDomain.setUsername(sipClient.getLabel());
-        sipClientDomain.setPhoneNumber(sipClient.getPhoneNr());
-        sipClientDomain.setCompany(sipClient.getCompany().getName());
-        sipClientDomain.setSecret(sipClient.getSecret());
-        sipClientDomain.setId(sipClient.getId());
-
-        return sipClientDomain;
     }
 
     /**
@@ -69,7 +46,7 @@ public class ConfigServiceImpl implements ConfigServiceInterface {
     @Override
     public ByteArrayOutputStream getAsteriskConfiguration() {
         List<SipClientConfigurationInterface> sipClientList = sipClientConfigAdapter.getSipClientList();
-        List<DialPlanContextConfigurationInterface> contexts = getDialPlanContexts();
+        List<? extends DialPlanContextConfigurationInterface> contexts = dialPlanConfigAdapter.getDialPlan();
 
         String sipClientConf = ConfigWriter.generateSipClientConfiguration(sipClientList);
         String dialPlanConf = ConfigWriter.generateDialPlanConfiguration(contexts);
@@ -77,20 +54,6 @@ public class ConfigServiceImpl implements ConfigServiceInterface {
         ConfigZipWriter configZipWriter = new ConfigZipWriter(sipClientConf, dialPlanConf);
 
         return configZipWriter.writeConfigurationZipFile();
-    }
-
-    private List<DialPlanContextConfigurationInterface> getDialPlanContexts() {
-        List<DialPlanContext> dialPlanContextList = dialPlanConfigAdapter.getDialPlanContextList();
-
-        List<Company> companyList = dialPlanConfigAdapter.getCompanyDomainList();
-        //TODO probably it would be better if this call is in the domain, and a list of companies is passed to the domain
-        List<DialPlanContext> defaultContexts = ContextGenerator.getDefaultContexts(companyList);
-
-        List<DialPlanContextConfigurationInterface> contexts = new ArrayList<>();
-        contexts.addAll(dialPlanContextList);
-        contexts.addAll(defaultContexts);
-
-        return contexts;
     }
 
 }
