+++ src/main/java/ch/zhaw/psit4/services/implementation/adapters/DialPlanConfigAdapter.java
@@ -1,17 +1,15 @@
 package ch.zhaw.psit4.services.implementation.adapters;
 
-import ch.zhaw.psit4.data.jpa.repositories.CompanyRepository;
 import ch.zhaw.psit4.data.jpa.repositories.SipClientRepository;
-import ch.zhaw.psit4.domain.beans.Company;
-import ch.zhaw.psit4.domain.beans.DialPlanContext;
+import ch.zhaw.psit4.domain.CompanyDialPlanBuilder;
 import ch.zhaw.psit4.domain.beans.SipClient;
-import ch.zhaw.psit4.services.implementation.ConfigServiceImpl;
+import ch.zhaw.psit4.domain.interfaces.DialPlanContextConfigurationInterface;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
+import java.util.stream.Collectors;
+import java.util.stream.StreamSupport;
 
 /**
  * Helps to gather all dial plan data from the data storage and converts it to the domain specific objects.
@@ -20,47 +18,22 @@ import java.util.List;
  */
 public class DialPlanConfigAdapter {
     private static final Logger LOGGER = LoggerFactory.getLogger(DialPlanConfigAdapter.class);
-    private CompanyRepository companyRepository;
     private SipClientRepository sipClientRepository;
 
-    public DialPlanConfigAdapter(SipClientRepository sipClientRepository, CompanyRepository companyRepository) {
-        this.companyRepository = companyRepository;
+    public DialPlanConfigAdapter(SipClientRepository sipClientRepository) {
         this.sipClientRepository = sipClientRepository;
     }
 
-    /**
-     * Converts the dial plans of the data storage into domain specific dial plans.
-     *
-     * @return domain specific dial plans
-     */
-    public List<DialPlanContext> getDialPlanContextList() {
-        // TODO convert the data
-        return Collections.emptyList();
-    }
 
-    /**
-     * Converts the companies of the data storage into domain specific companies.
-     *
-     * @return the domain specific companies
-     */
-    public List<Company> getCompanyDomainList() {
-        List<Company> companyList = new ArrayList<>();
-        for (ch.zhaw.psit4.data.jpa.entities.Company company : companyRepository.findAll()) {
-            Company companyDomain = new Company();
-            companyDomain.setName((company.getName()));
-            companyDomain.setSipClientList(getSipClientsByCompany(company));
-            companyList.add(companyDomain);
-        }
-        return companyList;
-    }
+    public List<? extends DialPlanContextConfigurationInterface> getDialPlan() {
+        CompanyDialPlanBuilder companyDialPlanBuilder = new CompanyDialPlanBuilder();
 
-    private List<SipClient> getSipClientsByCompany(ch.zhaw.psit4.data.jpa.entities.Company company) {
-        List<SipClient> sipClientList = new ArrayList<>();
-        for (ch.zhaw.psit4.data.jpa.entities.SipClient sipClient : sipClientRepository.findByCompany(company)) {
-            LOGGER.info("Add sipClient " + sipClient.getLabel());
-            SipClient sipClientDomain = ConfigServiceImpl.sipClientEntityToSipClient(sipClient);
-            sipClientList.add(sipClientDomain);
-        }
-        return sipClientList;
+        List<SipClient> sipClients =
+                StreamSupport.stream(sipClientRepository.findAll().spliterator(), false)
+                        .map(SipClientConfigAdapter::sipClientEntityToSipClient)
+                        .collect(Collectors.toList());
+
+        companyDialPlanBuilder.perCompanyDialExtensions(sipClients);
+        return companyDialPlanBuilder.build();
     }
 }
